requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
cloudinary==1.44.1
colorama==0.4.6
distro==1.9.0
Django==6.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
openai==2.24.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.7.5
python-decouple==3.8
python-docx==1.2.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
